pyproject.toml
[project]
name = "ctf-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "cryptography>=45.0.6",
    "gmpy2>=2.2.1",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "pillow-heif>=1.1.0",
    "pillow>=11.3.0",
    "pwntools>=4.14.1",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1,<3",
    "black>=25.1.0,<26",
    "deptry>=0.23.0,<0.24",
    "flake8>=7.2.0,<8",
    "isort>=6.0.1,<7",
    "mypy>=1.15.0,<2",
    "pre-commit==4.2.0",
    "pytest>=8.2.3,<9",
    "pytest-asyncio>=0.23.7,<0.24",
    "pytest-cov>=5.0.0,<6",
    "ruff>=0.11.8,<0.12",
    "taskiq[reload]>=0.11.17,<0.12",
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py312']
